+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased]
 
-tbd
+### Fixed ("fix")
+
+* fix(redfish.py): make lib more robust and check the sensor state against the user's thresholds first
+
 
 
 ## [v2.3.0] - 2025-06-20